Move generated client & essential tooling
The integration testing environment is largely
the same, as well as the script for generating the
client. Although there are not currently tests
for the OIDC & LDAP authenticators included in the
client I left them in as a part of the integration
testing infrastructure, it is zero cost to leave them
in & we will definitely want to test against them in
the future.